+++ example/ttt/game.dart
@@ -1,7 +1,7 @@
 library icacoe;
 
 import 'dart:html';
-import 'package:boarding/boarding_model.dart';
+import 'package:boarding/grids.dart';
 import 'package:boarding/boarding.dart';
 
 part 'model/square_grid.dart';