+++ lib/art_pen.dart
@@ -1,0 +1,27 @@
+library art_pen;
+
+import 'dart:math';
+
+import 'package:boarding/util.dart' as util;
+import 'package:dartling/dartling.dart';
+
+part 'model/art_pen/art/pen/json/data.dart';
+part 'model/art_pen/art/pen/json/model.dart';
+
+part 'model/art_pen/art/pen/init.dart';
+part 'model/art_pen/art/pen/segments.dart';
+part 'model/art_pen/art/pen/lines.dart';
+
+part 'model/art_pen/gen/art/pen/entries.dart';
+part 'model/art_pen/gen/art/pen/segments.dart';
+part 'model/art_pen/gen/art/pen/lines.dart';
+part 'model/art_pen/gen/art/models.dart';
+part 'model/art_pen/gen/art/repository.dart';
+
+// added by hand
+part 'model/art_pen/art/pen/pen.dart';
+part 'model/art_pen/art/pen/examples.dart';
+part 'model/art_pen/art/pen/programs.dart';
+
+
+