+++ lib/grids.dart
@@ -1,0 +1,6 @@
+library grids;
+
+part 'model/grids/cell.dart';
+part 'model/grids/grid.dart';
+
+