+++ lib/model/grids/cell.dart
@@ -1,4 +1,4 @@
-part of boarding_model;
+part of grids;
 
 class Cell {
   int row, column;