+++ lib/pieces.dart
@@ -1,0 +1,5 @@
+library pieces;
+
+import 'package:boarding/util.dart';
+
+part 'model/pieces/pieces.dart';