+++ lib/util.dart
@@ -1,0 +1,6 @@
+library util;
+
+import 'dart:math';
+
+part 'util/color.dart';
+part 'util/random.dart';